requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
PyMySQL==1.1.0
PyYAML==6.0.1
requests==2.31.0
APScheduler==3.10.4
cryptography==41.0.7
